seperate DBConnector: insert/find&delete
